set prescirption line state
